manosBackend/requirements.txt
coreapi==2.3.0
coreschema==0.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.1
django-rest-swagger==2.1.2
django-toolbelt==0.0.1
djangorestframework==3.6.3
drfdocs==0.0.11
gunicorn==19.7.1
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
openapi-codec==1.3.1
passlib==1.7.1
psycopg2==2.7.1
pytz==2017.2
requests==2.14.2
simplejson==3.10.0
static3==0.7.0
uritemplate==3.0.0
